Caso_Practico_Final_Maestria/Farmacia/requirements.txt
Flask
pandas
numpy
scikit-learn
matplotlib
tensorflow
nltk
requests
llama-cpp-python
pytesseract
Pillow
opencv-python
